+++ packages/measure/src/index.ts
@@ -1,5 +1,5 @@
-export { configure, resetToDefault } from './config';
-export { measurePerformance } from './measure-renders';
+export { configure, resetToDefaults } from './config';
+export { measureRenders, measurePerformance } from './measure-renders';
 export { measureFunction } from './measure-function';
-export type { MeasureOptions } from './measure-renders';
+export type { MeasureRendersOptions } from './measure-renders';
 export type { MeasureFunctionOptions } from './measure-function';