+++ packages/reassure/src/index.ts
@@ -1,4 +1,10 @@
-export { measurePerformance, measureFunction, configure, resetToDefault } from '@callstack/reassure-measure';
+export {
+  measureRenders,
+  measureFunction,
+  configure,
+  resetToDefaults,
+  measurePerformance,
+} from '@callstack/reassure-measure';
 export { dangerReassure } from '@callstack/reassure-danger';
 
-export type { MeasureOptions, MeasureFunctionOptions } from '@callstack/reassure-measure';
+export type { MeasureRendersOptions, MeasureFunctionOptions } from '@callstack/reassure-measure';