+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @callstack/reassure-cli
 
+## 1.0.0-beta
+
+### Major Changes
+
+- v1.0
+
+### Patch Changes
+
 ## 0.11.0
 
 ### Patch Changes