+++ packages/measure/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @callstack/reassure-measure
 
+## 1.0.0-beta
+
+### Major Changes
+
+- v1.0
+
+### Minor Changes
+
+- 4352279: - Rename `measurePerformance` to `measureRenders`.
+
+### Patch Changes
+
+- Updated dependencies [f9116d0]
+  - @callstack/reassure-logger@1.0.0-beta
+
 ## 0.11.0
 
 ### Minor Changes