+++ test-apps/native/CHANGELOG.md
@@ -1,5 +1,15 @@
 # example-native
 
+## 1.0.0-beta
+
+### Major Changes
+
+- v1.0
+
+### Minor Changes
+
+- 4352279: - Rename `measurePerformance` to `measureRenders`, `resetToDefault` to `resetToDefaults`.
+
 ## 0.1.0
 
 ### Minor Changes